% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doGSEAft.R
\name{doGSEAft}
\alias{doGSEAft}
\title{GSEA with Fisher's Exact Test}
\usage{
doGSEAft(selEID.v, ptw.ls, allEID.v, ncores = 4, minN = 5, adjPVth = 0.05)
}
\arguments{
\item{selEID.v}{A vector of selected Entrez Gene ID.}

\item{ptw.ls}{Lists of Gene EntrezID in each pathway of interest. You can get the 8567 biological terms from Molecular Signatures Database by `data("MSigDB-28Feb14-data")`.}

\item{allEID.v}{A vector of the universal set of Entrez Gene ID which you select genes from.}

\item{ncores}{Number of cores used for parallel running. (default = 4)}

\item{minN}{For each pathway, the minium number of genes(i.e. available in the ranked gene list) to conduct GSEA. If less than this value, the p value of this pathway would be set 1. (default = 5)}

\item{adjPVth}{Adjusted p value threshold to infer a pathway to be significantly enriched or not. P value was derived from Wilcoxon rank sum test and adjusted with BH method. (default = 0.05)}
}
\value{
\item{Rank(P)}{A matrix showing enriched pathways ranked by adjusted Fisher's Exact Test p values.
"nREP" is the number of genes in the pathway,
"nOVL" is the number of selected genes in the pathway,
"OR" is the odds ratio of Fisher's Exact Test,
"P" is the p value of Fisher's Exact Test,
"adjP" is the adjusted p value of Fisher's Exact Test (method='BH'),
"Genes" is all the selected genes in the pathway.}

\item{Rank(OR)}{A matrix showing enriched pathways ranked by odds ratio. The columns are samely defined as in Rank(P).}
}
\description{
Perform GSEA with Fisher's Exact Test to a group of selected genes
}
\details{
GSEA with Fisher's Exact Test is an extended enrichement test for user specified genes that doesn't require for ranks.
}
\examples{
# topGSEAft.lm <- doGSEAft(selEID.v = sigEID.ls$selEID, ptw.ls = listEZ.lv, allEID.v = names(mapEIDto450k.lv), ncores = 1, adjPVth = 0.05)

# Details can be found in tutorial

}
\references{
Dong D, Tian Y, Zheng SC, Teschendorff AE.
\emph{ebGSEA: an improved Gene Set Enrichment Analysis method for Epigenome-Wide-Association Studies.}
BMC Bioinformatics (2019) 35(18):3514-3516.
doi:\href{https://doi.org/10.1093/bioinformatics/btz073}{
10.1093/bioinformatics/btz073}.
}
\author{
Andrew E. Teschendorff, Tianyu Zhu
}
