% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doGSEAwt.R
\name{doGSEAwt}
\alias{doGSEAwt}
\title{GSEA with Wilcoxon Rank Sum Test and the Known-Population Median Test}
\usage{
doGSEAwt(rankEID.m, ptw.ls, ncores = 4, minN = 5, adjPVth = 0.05)
}
\arguments{
\item{rankEID.m}{The resulted matrix from doGT function, with genes by row and ranked by statistics from global test. Rownames of the matrix should be gene EntrezID.}

\item{ptw.ls}{Lists of Gene EntrezID in each pathway of interest. You can get the 8567 biological terms from Molecular Signatures Database by `data("MSigDB-28Feb14-data")`.}

\item{ncores}{Number of cores used for parallel running. (default = 4)}

\item{minN}{For each pathway, the minium number of genes(i.e. available in the ranked gene list) to conduct GSEA. If less than this value, the p value of this pathway would be set 1. (default = 5)}

\item{adjPVth}{Adjusted p value threshold to infer a pathway to be significantly enriched or not. P value was derived from Wilcoxon rank sum test and adjusted with BH method. (default = 0.05)}
}
\value{
\item{Rank(P)}{A matrix showing enriched pathways ranked by adjusted Wilcox test p values.
"nREP" is the number of mapped genes in the pathway,
"AUC" is the Area under curve of wilcox test,
"P(WT)" is the p-value of wilcox test,
"P(KPMT)" is the p-value of the Known-Population Median Test,
"adjP" is the adjusted p-value of wilcox test.}

\item{Rank(AUC)}{A matrix showing enriched pathways ranked by AUC. The columns are defined samely as Rank(P).}

\item{Genestat}{Lists of gene symbols in each enriched pathway. Each object contains the statistic and p-value from global test of each gene.}
}
\description{
Perform GSEA with wilcoxon rank sum test and known-population test using the ranked gene list from global test.
}
\details{
GSEA with Wilcoxon Rank Sum Test and the Known-Population Median Test is the second step of ebGSEA algorithm. Once the ranks of genes were derived from \emph{doGT}, enrichment of biological terms can be performed using either a standard one-tailed Wilcoxon rank sum test (WT), or a recently introduced more powerful version called Known-Population Median Test (KPMT). A group of enriched pathways can then be defined over adjusted pvalue from Wilcoxon rank sum test.
}
\examples{
# data("MSigDB-28Feb14-data")
# data("sgtm")
# topGSEA.lm <- doGSEAwt(rankEID.m = sgt.m, ptw.ls = listEZ.lv, ncores = 10, minN = 5, adjPVth = 0.05)

}
\references{
Dong D, Tian Y, Zheng SC, Teschendorff AE.
\emph{ebGSEA: an improved Gene Set Enrichment Analysis method for Epigenome-Wide-Association Studies.}
BMC Bioinformatics (2019) 35(18):3514-3516.
doi:\href{https://doi.org/10.1093/bioinformatics/btz073}{
10.1093/bioinformatics/btz073}.
}
\author{
Andrew E. Teschendorff, Tianyu Zhu
}
