% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doGT.R
\name{doGT}
\alias{doGT}
\title{Empirical Bayes Global Test}
\usage{
doGT(pheno.v, data.m, model = c("linear"), array = c("450k", "850k"), ncores = 4)
}
\arguments{
\item{pheno.v}{A vector of phenotype information, must be matched to columns of the input beta matrix.}

\item{data.m}{A matrix of beta values with probes by row and samples by column. Missing values shoud be excluded.}

\item{model}{The regression model for global test. Default is "linear".}

\item{array}{Array type for the input data. "450k" for Illumina HumanMethylation450 data and "850k" for Illumina MethylationEPIC data.}

\item{ncores}{Number of cores used for parallel running. (default = 4)}
}
\value{
A matrix with genes in row ranked by statistic from global test.
}
\description{
Function to assess the overall level of differential methylation using all the probes mapping to a gene.
}
\details{
Global test is the first step of ebGSEA algorithm. ebGSEA ranks genes according to their overall level of differential methylation by adapting the global test from \emph{Geoman et al(2006)}, which can be interpreted as an empirical Bayes generalized regression model. The global test evaluates whether DNA methylation patterns of CpGs mapping to a given gene \emph{g} differ significantly between two phenotypes.
}
\examples{
# sgt.m <- doGT(pheno.v, data.m, array = c("450k"), ncores = 10)
# topGSEA.lm <- doGSEAwt(rankEID.m = sgt.m, ptw.ls = listEZ.lv, ncores = 10, minN = 5, adjPVth = 0.05)


}
\references{
Dong D, Tian Y, Zheng SC, Teschendorff AE.
\emph{ebGSEA: an improved Gene Set Enrichment Analysis method for Epigenome-Wide-Association Studies.}
BMC Bioinformatics (2019) 35(18):3514-3516.
doi:\href{https://doi.org/10.1093/bioinformatics/btz073}{
10.1093/bioinformatics/btz073}.
}
\author{
Andrew E. Teschendorff, Tianyu Zhu
}
