% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dualmap450kEID}
\alias{dualmap450kEID}
\title{Illumina HM450k probes annotation to Entrez Gene ID}
\format{
A list with 485577 items and a list with 18649 items
}
\usage{
data("dualmap450kEID")
}
\description{
This annotation file is derived from Illumina HM450k annotation
}
\details{
\itemize{
  \item map450ktoEID.lv : A list mapping 450k probes to Entrez Gene ID
  \item mapEIDto450k.lv : A list mapping Entrez Gene ID to 450k probes
}
}
\references{
Dong D, Tian Y, Zheng SC, Teschendorff AE.
\emph{ebGSEA: an improved Gene Set Enrichment Analysis method for Epigenome-Wide-Association Studies.}
BMC Bioinformatics (2019) 35(18):3514-3516.
doi:\href{https://doi.org/10.1093/bioinformatics/btz073}{
10.1093/bioinformatics/btz073}.
}
\keyword{annotation}
