% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dualmap850kEID}
\alias{dualmap850kEID}
\title{Illumina EPIC probes annotation to Entrez Gene ID}
\format{
A list with 867531 items and a list with 24357 items
}
\usage{
data("dualmap850kEID")
}
\description{
This annotation file is derived from Illumina annotation file
}
\details{
\itemize{
  \item map850ktoEID.lv : A list mapping 850k probes to Entrez Gene ID
  \item mapEIDto850k.lv : A list mapping Entrez Gene ID to 850k probes
}
}
\references{
Dong D, Tian Y, Zheng SC, Teschendorff AE.
\emph{ebGSEA: an improved Gene Set Enrichment Analysis method for Epigenome-Wide-Association Studies.}
BMC Bioinformatics (2019) 35(18):3514-3516.
doi:\href{https://doi.org/10.1093/bioinformatics/btz073}{
10.1093/bioinformatics/btz073}.
}
\keyword{annotation}
