\docType{data}
\name{ebgseaDATA}
\alias{ebgseaDATA}
\title{Sample DNAm data from a buccal swab study}
\format{
Two objects: a matrix containing the DNAm data and a matched phenotype
vector containing the smoking-pack-year information
}
\usage{
data("ebgseaDATA")
}
\description{
A DNAm data matrix over 7933 CpGs and 325 samples, and a corresponding
vector with 325 entries.
}
\details{
\itemize{
  \item dataSMK.m : The DNAm data matrix defined over 7933 CpGs and 325 samples.
  \item phenoSMK.v : The vector containing the smoking information.
}
}
\references{
Dong D, Tian Y, Zheng SC, Teschendorff AE.
\emph{ebGSEA: an improved Gene Set Enrichment Analysis method for Epigenome-Wide-Association Studies.}
BMC Bioinformatics (2019) 35(18):3514-3516.
doi:\href{https://doi.org/10.1093/bioinformatics/btz073}{
10.1093/bioinformatics/btz073}.
}
\keyword{tutorial}
