% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MSigDB-28Feb14-data}
\alias{MSigDB-28Feb14-data}
\title{Biological terms from Molecular Signatures Database 28Feb14}
\format{
Two lists with 8567 items and a vector with 8567 items
}
\usage{
data("MSigDB-28Feb14-data")
}
\description{
The 8567 biological terms from \href{https://www.gsea-msigdb.org/gsea/msigdb/}{Molecular Signatures Database} 28Feb14
}
\details{
\itemize{
  \item listEZ.lv : Gene sets in NCBI (Entrez) Gene IDs for each biological term
  \item listG.lv : Gene stes in gene symbols for each biological term
  \item listclassALL.lv : The type of biological term defined by MSiDB
}
}
\references{
Dong D, Tian Y, Zheng SC, Teschendorff AE.
\emph{ebGSEA: an improved Gene Set Enrichment Analysis method for Epigenome-Wide-Association Studies.}
BMC Bioinformatics (2019) 35(18):3514-3516.
doi:\href{https://doi.org/10.1093/bioinformatics/btz073}{
10.1093/bioinformatics/btz073}.

Subramanian A, Tamayo P, Mootha VK, et al.
\emph{Gene set enrichment analysis: A knowledge-based approach for interpreting genome-wide expression profiles.}
Proc Natl Acad Sci U S A (2005) 102(43):15545‐15550.
doi:\href{https://doi.org/10.1073/pnas.0506580102}{
10.1073/pnas.0506580102}.
}
\keyword{biological}
\keyword{terms}
