% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SampleCpG}
\alias{SampleCpG}
\title{Sample CpGs for fisher's exact test in Tutorial}
\format{
A vector with length 40626 and a vector with length 484272
}
\usage{
data("SampleCpG")
}
\description{
The differentialy methylated cytosines associated with smoking pack-years identified from buccal swab dataset
}
\details{
\itemize{
  \item sampleCpG.v : The pre-selected differentially methylated cytosines
  \item allCpG.v : All the CpGs from buccal swab dataset
}
}
\references{
Dong D, Tian Y, Zheng SC, Teschendorff AE.
\emph{ebGSEA: an improved Gene Set Enrichment Analysis method for Epigenome-Wide-Association Studies.}
BMC Bioinformatics (2019) 35(18):3514-3516.
doi:\href{https://doi.org/10.1093/bioinformatics/btz073}{
10.1093/bioinformatics/btz073}.
}
\keyword{tutorial}
