% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selEIDfromSelCpG.R
\name{selEIDfromSelCpG}
\alias{selEIDfromSelCpG}
\title{Select significant genes from a selected set of CpGs}
\usage{
selEIDfromSelCpG(selCpG.v, allCpG.v, pvth = 0.3/length(selCpG.v), array = c("450k","850k"))
}
\arguments{
\item{selCpG.v}{A vector of user selected CpGs.}

\item{allCpG.v}{A vector of all CpGs the user select the CpGs from.}

\item{pvth}{P-value threshold to infer the number of selected CpGs mapped to a gene is significant or not in a binomial test. (default = 0.3/length(selCpG.v))}

\item{array}{Array type for the input CpGs. "450k" for Illumina HumanMethylation450 data and "850k" for Illumina MethylationEPIC data.}
}
\value{
\item{selEID}{A vector of significant genes in Entrez Gene ID.}

\item{selPV}{P-values of significant genes.}

\item{allPV}{P-values of all mapped genes.}
}
\description{
Select significant genes from a selected set of CpGs with binomial test.
}
\details{
This function maps user specified CpGs to genes and return genes whose number of mapped CpGs is significant in binomial test.
}
\examples{
# data("SampleCpG")
# sigEID.ls <- selEIDfromSelCpG(selCpG.v = sampleCpG.v, allCpG.v = allCpG.v, array = "450k")

}
\references{
Dong D, Tian Y, Zheng SC, Teschendorff AE.
\emph{ebGSEA: an improved Gene Set Enrichment Analysis method for Epigenome-Wide-Association Studies.}
BMC Bioinformatics (2019) 35(18):3514-3516.
doi:\href{https://doi.org/10.1093/bioinformatics/btz073}{
10.1093/bioinformatics/btz073}.
}
\author{
Andrew E. Teschendorff, Tianyu Zhu
}
